Microsoft tp omega (#1527)
* Accessibility & refactoring (#1468)

* Target word count 4.1.2.0 (#1469)

* CRQ-37360

* Fixed Object reference not set to an instance of an object

* SDLCOM-5549 (#1470)

* Data protection suite sdlcom 5329 (#1453)

* SDLCOM-5327 Tell Me Redirect & Refinement

* Data Protection Suite 3.1.5.0

* Updated tellme

* Removed unsed package

* Updated TellMe

* SDLCOM-5327

* U

* SDLCOM-5564

* SDLCOM-5657
SDLCOM-5656
SDLCOM-5564

* Fixed Studio paths

+++ TargetWordCount/Sdl.Community.TargetWordCount/ReportBuilder.cs
@@ -99,7 +99,7 @@ namespace Sdl.Community.TargetWordCount
 			}
 
 			var amount = count * rate;
-			int output = 0;
+			float output = 0;
 			int totalChar = 0;
 			if (settings.UseLineCount)
 			{
@@ -156,11 +156,11 @@ namespace Sdl.Community.TargetWordCount
 			}
 		}
 
-		private static int CalculateLineCount(RateType type, IWordCountBatchTaskSettings settings, CountData countData, CountTotal total, ref int output, ref int totalChar)
+		private static int CalculateLineCount(RateType type, IWordCountBatchTaskSettings settings, CountData countData, CountTotal total, ref float output, ref int totalChar)
 		{
 			if (!string.IsNullOrWhiteSpace(settings.CharactersPerLine))
 			{
-				if (int.TryParse(settings.CharactersPerLine, out output))
+				if (float.TryParse(settings.CharactersPerLine, out output))
 				{
 					if (type == RateType.Locked && settings.ReportLockedSeperately)
 					{
